Fix paul methods

+++ migrations/sql/init-scheme.sql
@@ -42,7 +42,7 @@ CREATE TABLE Card (
 
 CREATE TABLE MaxStatus(
     card_id INT REFERENCES Card(card_id),
-    max_status INT
+    max_status BIGINT
 );
 
 CREATE TABLE Cleaning (
